Add timestamp segment on client side on channel
Checks if the previous and current message have 2
hours difference. It behaves like timestamp segm-
ent in existing message. Because we have no way
to put it realtime, this check and insert is hap-
pening on the client side on channel.

Data timestamp attribute also added on messages
to retrieve raw timestamps.